fix(src/riak/config.go,src/riak/action.go): move config into riak proper
