+++ src/riak/config.go
@@ -1,4 +1,4 @@
-package config
+package riak
 
 import (
 	"github.com/kelseyhightower/envconfig"
@@ -8,15 +8,15 @@ const (
 	AppName = "deis-riak"
 )
 
-type Config struct {
+type config struct {
 	RiakMaster            bool   `envconfig:"RIAK_MASTER" default:"false"`
 	ClusterServerHTTPPort int    `envconfig:"CLUSTER_SERVER_HTTP_PORT" default:"8080"`
 	ClusterServerHost     string `envconfig:"DEIS_RIAK_CLUSTER_SERVICE_HOST" required:"true"`
 	ClusterServerPort     int    `envconfig:"DEIS_RIAK_CLUSTER_SERVICE_PORT" required:"true"`
 }
 
-func Get() (*Config, error) {
-	var ret Config
+func getConfig() (*config, error) {
+	var ret config
 	if err := envconfig.Process(AppName, &ret); err != nil {
 		return nil, err
 	}
